requirements.txt
streamlit
tensorflow-cpu
numpy
pillow
matplotlib
scikit-learn
pandas